nit: restructure code
