+++ crates/vertex_ai/src/data_store/error.rs
@@ -1,0 +1,2 @@
+#[derive(Debug, thiserror::Error)]
+pub enum Error {}
