+++ crates/vertex_ai/src/data_store/mod.rs
@@ -1,1 +1,2 @@
+mod data_store;
 mod error;
